requirements.txt
django
pip install git+git://github.com/dcramer/django-devserver#egg=django-devserver
-e git+https://github.com/chriscauley/django-registration.git#egg=django-registration
django_compressor
django-localflavor
djangorestframework
pip install djangorestframework-jwt
BeautifulSoup==3.2.1
psycopg2
django-suit
django-contrib-comments
ezdxf
simplejson

#image stuff
-e git+https://github.com/chriscauley/sorl-thumbnail.git#egg=sorl-thumbnail
-e git+https://github.com/chriscauley/Django-Crop-Override-Field.git#egg=django-crop-override-field
pillow==2.1.0

django-mptt==0.5.4
-e git+https://github.com/chriscauley/django-riot-comments.git#egg=django-riot-comments.git

django-redis-cache
Markdown
django-timezones==0.2
pytz>=2011n
Pygments
requests
-e git://github.com/durden/frappy.git#egg=frappy

django-tagging==0.3.6
django-next-please
icalendar
python-social-auth

-e git+https://github.com/chriscauley/django-shop.git#egg=django-shop
-e git+https://github.com/chriscauley/django-wmd-editor.git#egg=django_wmd_editor
-e git+ttps://github.com/chriscauley/lablackey.git#egg=lablackey
django_paypal==0.2.5
django-colorful
django-jsignature
